cleaning code, add crontroller for park capacity (getAllParksDay)
