+++ src/routes/parks.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const { body, validationResult } = require('express-validator');
+const { parksByDay } = require('../service/parksByDay');
 
 const router = express.Router();
 
@@ -22,7 +23,7 @@ module.exports = (storage) => {
         return res.status(400).json({ errors: errors.array() });
       }
       const { numOfGuests, date } = req.body;
-      const parks = await storage.parkByDay({ numOfGuests, date });
+      const parks = await parksByDay(storage, { numOfGuests, date });
       return res.json(parks);
     } catch (err) {
       return res.status(400).json({ error: err });
